+++ conveyor/src/main/java/com/aryunin/conveyor/service/CreditService.java
@@ -9,14 +9,14 @@ import java.util.List;
 // fixme check for division by null
 public interface CreditService {
     BigDecimal getRate(boolean isInsuranceEnabled, boolean isSalaryClient);
-    BigDecimal getMonthlyPayment(BigDecimal totalAmount, BigDecimal rate, Integer term);
-    BigDecimal getInsuranceAmount(BigDecimal amount);
+    BigDecimal getMonthlyPayment(BigDecimal amount, BigDecimal rate, Integer term);
     BigDecimal getTotalAmount(BigDecimal amount, boolean isInsuranceEnabled);
     BigDecimal getPSK(BigDecimal amount, BigDecimal monthlyPayment, Integer term);
     List<PaymentScheduleElement> getPaymentSchedule(
             LocalDate startDate,
-            BigDecimal totalAmount,
+            BigDecimal amount,
             BigDecimal rate,
             Integer term
     );
+
 }
