+++ conveyor/src/main/java/com/aryunin/conveyor/service/calculation/CalculationServiceImpl.java
@@ -1,7 +1,8 @@
-package com.aryunin.conveyor.service;
+package com.aryunin.conveyor.service.calculation;
 
 import com.aryunin.conveyor.dto.CreditDTO;
 import com.aryunin.conveyor.dto.ScoringDataDTO;
+import com.aryunin.conveyor.service.CreditService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -16,6 +17,7 @@ public class CalculationServiceImpl implements CalculationService{
     @Override
     public CreditDTO getCredit(ScoringDataDTO data) {
         var rate = creditService.getRate(data.getIsInsuranceEnabled(), data.getIsSalaryClient());
+
         rate = scoringService.scoreRate(rate, data);
 
         var totalAmount = creditService.getTotalAmount(data.getAmount(), data.getIsInsuranceEnabled());
